[REF] mail: move jumpUnread to discuss folder
This is channel specific code which should be in discuss folder.

Part of task-3265211
