+++ addons/mail/static/src/discuss/core/common/@types/models.d.ts
@@ -43,10 +43,13 @@ declare module "models" {
         readonly lastSelfMessageSeenByEveryone: Message,
         member_count: number | undefined,
         readonly membersThatCanSeen: ChannelMember[],
+        newMessageBannerText: string,
+        onClickUnreadMessagesBanner(): void,
         onlineMembers: ChannelMember[],
         offlineMembers: ChannelMember[],
         otherTypingMembers: ChannelMember[],
         selfMember: ChannelMember,
+        showUnreadBanner(): boolean,
         typingMembers: ChannelMember[],
         readonly hasMemberList: boolean,
         readonly notifyOnleave: boolean,
