+++ addons/mail/static/src/discuss/core/common/thread_patch.js
@@ -1,11 +1,38 @@
 import { Thread } from "@mail/core/common/thread";
 
-import { toRaw } from "@odoo/owl";
+import { useEffect, toRaw } from "@odoo/owl";
 
+import { _t } from "@web/core/l10n/translation";
 import { patch } from "@web/core/utils/patch";
 
-/** @type {import("models").Thread} */
+/** @type {Thread} */
 const threadPatch = {
+    setup() {
+        super.setup(...arguments);
+        useEffect(
+            (loadNewer, mountedAndLoaded, unreadSynced) => {
+                if (
+                    loadNewer ||
+                    unreadSynced || // just marked as unread (local and server state are synced)
+                    !mountedAndLoaded ||
+                    !this.props.thread.selfMember ||
+                    !this.scrollableRef.el
+                ) {
+                    return;
+                }
+                const el = this.scrollableRef.el;
+                if (Math.abs(el.scrollTop + el.clientHeight - el.scrollHeight) <= 1) {
+                    this.props.thread.selfMember.hideUnreadBanner = true;
+                }
+            },
+            () => [
+                this.props.thread.loadNewer,
+                this.state.mountedAndLoaded,
+                this.props.thread.selfMember?.unreadSynced,
+                this.state.scrollTop,
+            ]
+        );
+    },
     /** @override */
     fetchMessages() {
         if (this.props.thread.selfMember && this.props.thread.scrollUnread) {
@@ -20,6 +47,19 @@ const threadPatch = {
         if (this.selfMember && !this.isDisplayed) {
             this.selfMember.syncUnread = true;
         }
+    },
+    get newMessageBannerText() {
+        if (this.props.thread.selfMember?.totalUnreadMessageCounter > 1) {
+            return _t("%s new messages", this.props.thread.selfMember.totalUnreadMessageCounter);
+        }
+        return _t("1 new message");
+    },
+    async onClickUnreadMessagesBanner() {
+        await this.props.thread.loadAround(this.props.thread.selfMember.localNewMessageSeparator);
+        this.messageHighlight?.highlightMessage(
+            this.props.thread.firstUnreadMessage,
+            this.props.thread
+        );
     },
 };
 patch(Thread.prototype, threadPatch);
